pyproject.toml
[project]
name = "music-site-scripts"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mutagen>=1.47",
    "boto3>=1.35",
    "google-genai>=1.0",
    "pillow>=11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
